projects/slack_rag_assistant/requirements.txt
slack_sdk==3.19.0
openai==0.27.0
faiss-cpu==1.7.2
sentence-transformers==2.2.2
PyYAML==6.0
python-dotenv==0.20.0
requests==2.28.1
pytest==7.1.2
